+++ src/irmin-mirage/git/irmin_mirage_git.ml
@@ -3,7 +3,7 @@ open Lwt.Infix
 module type S = sig
   include
     Irmin_git.S
-      with type Private.Sync.endpoint = Mimic.ctx * Smart_git.Endpoint.t
+      with type Private.Remote.endpoint = Mimic.ctx * Smart_git.Endpoint.t
 
   val remote :
     ?ctx:Mimic.ctx -> ?headers:Cohttp.Header.t -> string -> Irmin.remote
@@ -103,7 +103,7 @@ module KV_RO (G : Git.S) = struct
   end
 
   module S = KV (G) (Irmin.Contents.String)
-  module Sync = Irmin.Sync (S)
+  module Sync = Irmin.Sync.Make (S)
 
   let disconnect _ = Lwt.return_unit