+++ src/irmin-pack/ext.mli
@@ -42,7 +42,7 @@ module Make
        and type step = Path.step
        and type metadata = Metadata.t
        and type Key.step = Path.step
-       and type Private.Sync.endpoint = unit
+       and type Private.Remote.endpoint = unit
 
   include Store.S with type repo := repo